pyproject.toml
[project]
    name = "film-colour"
    version = "0.1.0"
    description = "Add your description here"
    readme = "README.md"
    requires-python = "==3.11.*"
    dependencies = [
        "black>=26.1.0",
        "colour-science>=0.4.7",
        "joblib>=1.5.3",
        "matplotlib>=3.10.8",
        "numpy>=2.4.2",
        "opencv-python>=4.13.0.92",
        "scipy>=1.17.0",
        "tqdm>=4.67.3",
        "uv>=0.10.2",
]
